Mostly cosmetic refactorings. Begin adding documentation to support new developers and better describe our functions.

+++ src/Mass.h
@@ -15,8 +15,8 @@ class Mass : public Particle
 {
 public:
 
-    // Constructor
-    Mass(Vector2D position, Vector2D velocity, float mass, float size, float r, float g, float b) :
+    /* Constructor */
+    Mass(Vector2D position, Vector2D velocity, float mass, float size, GLubyte r, GLubyte g, GLubyte b) :
       Particle(position, velocity, mass),
       //bbDirty(false),
       size(size),
@@ -26,28 +26,25 @@ public:
       b(b)
       {};
 
-    // Default constructor
+    /* Default constructor */
     Mass() : 
       Particle(Vector2D(0.0f,0.0f),Vector2D(0.0f,0.0f), 0),
       size(1.0f)
       { SetColor(255, 255, 255); }
 
-    // Destructor
+    /* Destructor */
     virtual ~Mass(){};
 
-    // Applies gravitational force from another gravitational
-    // body.
+    /* Applies gravitational force from another body */
     void ApplyGravityFrom(Particle& body, float timestep);
 
-    // Draws this mass to the screen
+    /* Draws this mass to the screen */
     void Draw();
 
-
-    // Sets the size
+    /* Sets the size */
     void SetSize(float newsize) {size = newsize;}
 
-
-    // Gets the size
+    /* Gets the size*/
     float GetSize() const { return size;}
 
     // Get the status of the AABB
@@ -56,7 +53,7 @@ public:
     // Set the status of the AABB
     //void SetBBDirty(bool dirty) { bbDirty = dirty;}
 
-    // Set this mass' color
+    /* Set this mass' color */
     void SetColor(GLubyte R, GLubyte G, GLubyte B) { r = R; g = G; b = B; }
 
 
@@ -67,13 +64,13 @@ private:
     // Not Dirty = AABB is current.
     //bool bbDirty;
 
-    // The size of this mass
+    /* The size of this mass */
 	float size;
 
     // This mass's bounding box
     //AABB boundingBox;
 
-    // Color values from 0-255
+    /*  Color values from 0-255 */
     GLubyte r, g, b;
 };
 #endif
